+++ src/helpers/get-balance.ts
@@ -1,0 +1,4 @@
+export default function getBalance(price: number, amount: number) {
+  const balance = price * amount;
+  return balance.toFixed(2);
+}
