+++ src/helpers/get-best-interval.ts
@@ -1,0 +1,50 @@
+export default function getBestInterval(
+  range:
+    | '1d'
+    | '5d'
+    | '1mo'
+    | '3mo'
+    | '6mo'
+    | '1y'
+    | '2y'
+    | '5y'
+    | '10y'
+    | 'ytd'
+    | 'max',
+):
+  | '1m'
+  | '2m'
+  | '5m'
+  | '15m'
+  | '30m'
+  | '60m'
+  | '90m'
+  | '1h'
+  | '1d'
+  | '5d'
+  | '1wk'
+  | '1mo'
+  | '3mo' {
+  switch (range) {
+    case '1d':
+      return '5m';
+    case '5d':
+      return '1h';
+    case '1mo':
+      return '1d';
+    case '3mo':
+      return '1d';
+    case '6mo':
+      return '1d';
+    case '1y':
+      return '5d';
+    case '2y':
+      return '1wk';
+    case '5y':
+      return '1mo';
+    case '10y':
+      return '3mo';
+    case 'ytd':
+      return '1d';
+  }
+}
